This allows you to manage symbol keys
Given a hash like

```puppet
$symbols = {
  ":bar" => {
    "baz" => "foo",
    "moo" => 1,
    ":cow" => "1",
  }
}
```

Prior to this commit, this type would manage a hash with literal keys
with the colon in them and it was impossible to manage a hash with
symbolic keys. This commit will convert a string like `":thing"` to a
symbol, like `:thing` when managing yaml files.

Fixes #2